Moved message classes to Models/Messages
